// src/app/interface/medicine.interface.ts
export interface Medicine {
    dosage_type: string;
    medicine_type: string;
    schedule_type: string;
    cess_percentage: number;
    mrp: number;
    packing_size: string;
    manufacturer_name: string;
    medicine_name: string;
    content: string;
    gst_percentage: number;
    size: number;
    price: number;
    available_for_patient: number;
    discontinued: number;
    medicine_id:number;
  }